% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIDEOD.R
\docType{data}
\name{chideod}
\alias{chideod}
\title{Descriptive data title}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{traditional}{}
\item{simplified}{}
\item{pinyin_tone}{}
\item{pinyin_tonenumber}{}
\item{pinyin_without_tone}{}
\item{ipa_toneletter}{}
\item{ipa_tonenumber}{}
\item{middle_chinese_baxter}{}
\item{middle_chinese_ipa}{}
\item{old_chinese_ipa}{}
\item{language_stage}{}
\item{data_source}{}
\item{traditional1}{}
\item{traditional2}{}
\item{traditional3}{}
\item{traditional4}{}
\item{simplified1}{}
\item{simplified2}{}
\item{simplified3}{}
\item{simplified4}{}
\item{sensory_imagery}{}
\item{definitions}{}
\item{orthograpic_variants}{}
\item{morphological_template}{}
\item{radical_support}{}
\item{character1_freq}{}
\item{character2_freq}{}
\item{character3_freq}{}
\item{character4_freq}{}
\item{character1_family_size}{}
\item{character2_family_size}{}
\item{character3_family_size}{}
\item{character4_family_size}{}
\item{character1_semantic_radical}{}
\item{character2_semantic_radical}{}
\item{character3_semantic_radical}{}
\item{character4_semantic_radical}{}
\item{character1_semantic_radical_freq}{}
\item{character2_semantic_radical_freq}{}
\item{character3_semantic_radical_freq}{}
\item{character4_semantic_radical_freq}{}
\item{character1_semantic_family_size}{}
\item{character2_semantic_family_size}{}
\item{character3_semantic_family_size}{}
\item{character4_semantic_family_size}{}
\item{character1_phonetic_component}{}
\item{character2_phonetic_component}{}
\item{character3_phonetic_component}{}
\item{character4_phonetic_component}{}
\item{character1_phonetic_component_freq}{}
\item{character2_phonetic_component_freq}{}
\item{character3_phonetic_component_freq}{}
\item{character4_phonetic_component_freq}{}
\item{character1_phonetic_family_size}{}
\item{character2_phonetic_family_size}{}
\item{character3_phonetic_family_size}{}
\item{character4_phonetic_family_size}{}
\item{interjection}{}
\item{note}{}
}}
\source{
The data comes from https://osf.io/kpwg.
}
\description{
Detailed description of the data
}
\seealso{
\link{CHIDEOD}
}
