% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIDEOD.R
\docType{package}
\name{CHIDEOR}
\alias{CHIDEOR}
\alias{CHIDEOR-package}
\title{CHIDEOD -- The Chinese Ideophone Database}
\description{
This package contains the Chinese Ideophone Database CHIDEOD.
}
\details{
CHIDEOD
A data package for CHIDEOD.

Use \code{data(package='CHIDEOR')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{chideod}
}
