% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIDEOD.R
\docType{data}
\name{chideod}
\alias{chideod}
\title{The Chinese Ideophone Database (CHIDEOD)}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{traditional}{}
\item{simplified}{}
\item{pinyintone}{}
\item{pinyinnum}{}
\item{pinyinnone}{}
\item{HYDCD}{}
\item{zdic}{}
\item{Kroll}{}
\item{MC}{}
\item{OC}{}
\item{radsup}{}
\item{S1}{}
\item{S2}{}
\item{S3}{}
\item{S4}{}
\item{T1}{}
\item{T2}{}
\item{T3}{}
\item{T4}{}
\item{morphology}{}
\item{S1.charfreq}{}
\item{S2.charfreq}{}
\item{S3.charfreq}{}
\item{S4.charfreq}{}
\item{S1.famfreq}{}
\item{S2.famfreq}{}
\item{S3.famfreq}{}
\item{S4.famfreq}{}
\item{S1.sem}{}
\item{S2.sem}{}
\item{S3.sem}{}
\item{S4.sem}{}
\item{S1.semfreq}{}
\item{S2.semfreq}{}
\item{S3.semfreq}{}
\item{S4.semfreq}{}
\item{S1.semfam}{}
\item{S2.semfam}{}
\item{S3.semfam}{}
\item{S4.semfam}{}
\item{S1.phon}{}
\item{S2.phon}{}
\item{S3.phon}{}
\item{S4.phon}{}
\item{S1.phonfreq}{}
\item{S2.phonfreq}{}
\item{S3.phonfreq}{}
\item{S4.phonfreq}{}
\item{S1.phonfam}{}
\item{S2.phonfam}{}
\item{S3.phonfam}{}
\item{S4.phonfam}{}
\item{sensorymod}{}
\item{variant}{}
\item{note}{}
\item{datasource}{}
\item{stage}{}
\item{ASBC.dp}{}
\item{ASBC.philosophy}{}
\item{ASBC.literature}{}
\item{ASBC.life}{}
\item{ASBC.society}{}
\item{ASBC.science}{}
\item{ASBC.art}{}
}}
\source{
The data comes from https://osf.io/kpwg.
}
\description{
Detailed description of the data
}
\seealso{
\link{CHIDEOD}
}
